+++ myfoodapp.Hub/Viewmodels/ProductionUnitViewModel.cs
@@ -19,9 +19,9 @@ namespace myfoodapp.Hub.Models
 
         public ProductionUnitTypeViewModel productionUnitType { get; set; }
  
-        public decimal locationLatitude { get; set; }
+        public double locationLatitude { get; set; }
 
-        public decimal locationLongitude { get; set; }
+        public double locationLongitude { get; set; }
         public String info { get; set; }
         public String version { get; set; }
         public List<Option> options { get; set; }
